rewrite new entity forms and new entity modal to allow connecting stuff from modal #36, also addresses #99
