--- app/containers/ActionNew/actions.js
@@ -1,16 +1,0 @@
-/*
- *
- * ActionNew actions
- *
- */
-
-import {
-  SAVE,
-} from './constants';
-
-export function save(data) {
-  return {
-    type: SAVE,
-    data,
-  };
-}
