+++ app/containers/ActionNew/constants.js
@@ -6,8 +6,6 @@
 import { fromJS } from 'immutable';
 import { API, ACTION_FIELDS } from 'themes/config';
 
-export const SAVE = 'impactoss/ActionNew/SAVE';
-
 export const DEPENDENCIES = [
   API.USER_ROLES,
   API.CATEGORIES,
