+++ app/containers/CategoryNew/CategoryNewForm.js
@@ -1,0 +1,347 @@
+/*
+ *
+ * CategoryNewForm
+ *
+ */
+
+import React from 'react';
+import PropTypes from 'prop-types';
+import { connect } from 'react-redux';
+import { actions as formActions } from 'react-redux-form/immutable';
+
+import { List } from 'immutable';
+
+import {
+  renderParentCategoryControl,
+  getTitleFormField,
+  getMarkdownFormField,
+  getFormField,
+  getCheckboxField,
+  getStatusField,
+  getShortTitleFormField,
+} from 'utils/forms';
+
+import { scrollToTop } from 'utils/scroll-to-component';
+import { hasNewErrorNEW } from 'utils/entity-form';
+
+import { getCheckedValuesFromOptions } from 'components/forms/MultiSelectControl';
+
+import { CONTENT_SINGLE, CONTENT_MODAL } from 'containers/App/constants';
+import { API, ROUTES, USER_ROLES } from 'themes/config';
+
+import appMessages from 'containers/App/messages';
+
+import {
+  loadEntitiesIfNeeded,
+  redirectIfNotPermitted,
+  updatePath,
+  updateEntityForm,
+  submitInvalid,
+  saveErrorDismiss,
+  openNewEntityModal,
+  newEntity,
+} from 'containers/App/actions';
+
+import {
+  selectReady,
+  selectReadyForAuthCheck,
+  selectIsUserAdmin,
+  selectTaxonomy,
+} from 'containers/App/selectors';
+
+import Content from 'components/Content';
+import ContentHeader from 'components/ContentHeader';
+
+import { getEntityTitle } from 'utils/entities';
+
+import FormWrapper from './FormWrapper';
+
+import {
+  selectParentOptions,
+  selectParentTaxonomy,
+} from './selectors';
+
+import messages from './messages';
+import { DEPENDENCIES, FORM_INITIAL } from './constants';
+
+
+export class CategoryNewForm extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
+  constructor(props) {
+    super(props);
+    this.scrollContainer = React.createRef();
+  }
+
+  UNSAFE_componentWillMount() {
+    this.props.loadEntitiesIfNeeded();
+    this.props.initialiseForm(FORM_INITIAL);
+  }
+
+  UNSAFE_componentWillReceiveProps(nextProps) {
+    // reload entities if invalidated
+    if (!nextProps.dataReady) {
+      this.props.loadEntitiesIfNeeded();
+    }
+    if (nextProps.authReady && !this.props.authReady) {
+      this.props.redirectIfNotPermitted();
+    }
+    if (hasNewErrorNEW(nextProps, this.props) && this.scrollContainer) {
+      scrollToTop(this.scrollContainer.current);
+    }
+  }
+
+  getHeaderMainFields = () => {
+    const { intl } = this.context;
+    const groups = [];
+    groups.push({ // fieldGroup
+      fields: [
+        getTitleFormField(intl.formatMessage),
+        getShortTitleFormField(intl.formatMessage),
+      ],
+    });
+    return groups;
+  };
+
+  getHeaderAsideFields = (taxonomy, parentOptions, parentTaxonomy) => {
+    const { intl } = this.context;
+    const groups = []; // fieldGroups
+    groups.push({
+      fields: [
+        getStatusField(intl.formatMessage),
+        getStatusField(intl.formatMessage, 'private'),
+      ],
+    });
+    if (taxonomy.getIn(['attributes', 'tags_users'])) {
+      groups.push({
+        fields: [
+          getCheckboxField(
+            intl.formatMessage,
+            'user_only',
+          ),
+        ],
+      });
+    }
+    if (parentOptions && parentTaxonomy) {
+      groups.push({
+        label: intl.formatMessage(appMessages.entities.taxonomies.parent),
+        icon: 'categories',
+        fields: [renderParentCategoryControl(
+          parentOptions,
+          getEntityTitle(parentTaxonomy),
+        )],
+      });
+    }
+    return groups;
+  }
+
+  getBodyMainFields = () => {
+    const { intl } = this.context;
+    const groups = [];
+    groups.push({
+      fields: [getMarkdownFormField(intl.formatMessage)],
+    });
+    return groups;
+  };
+
+  getBodyAsideFields = () => {
+    const { intl } = this.context;
+    const fields = []; // fieldGroups
+    fields.push({
+      fields: [
+        getFormField({
+          formatMessage: intl.formatMessage,
+          controlType: 'url',
+          attribute: 'url',
+        }),
+      ],
+    });
+    return fields;
+  }
+
+  /* eslint-disable react/destructuring-assignment */
+  getTaxTitle = (id) => this.context.intl.formatMessage(appMessages.entities.taxonomies[id].single);
+  /* eslint-enable react/destructuring-assignment */
+
+  render() {
+    const { intl } = this.context;
+    const {
+      taxonomy,
+      dataReady,
+      viewDomain,
+      parentOptions,
+      parentTaxonomy,
+      onErrorDismiss,
+      onServerErrorDismiss,
+      handleCancel,
+      handleSubmitRemote,
+      handleSubmit,
+      handleSubmitFail,
+      handleUpdate,
+      typeId,
+      formDataPath,
+      inModal,
+    } = this.props;
+    const { saveSending } = viewDomain.get('page').toJS();
+
+    let pageTitle = intl.formatMessage(messages.pageTitle);
+    if (taxonomy && taxonomy.get('attributes')) {
+      pageTitle = intl.formatMessage(messages.pageTitleTaxonomy, {
+        taxonomy: this.getTaxTitle(taxonomy.get('id')),
+      });
+    }
+
+    return (
+      <Content ref={this.scrollContainer} inModal={inModal}>
+        <ContentHeader
+          title={pageTitle}
+          type={inModal ? CONTENT_MODAL : CONTENT_SINGLE}
+          buttons={
+            dataReady ? [{
+              type: 'cancel',
+              onClick: () => handleCancel(typeId),
+            },
+            {
+              type: 'save',
+              disabled: saveSending,
+              onClick: () => handleSubmitRemote(formDataPath),
+            }] : null
+          }
+        />
+        <FormWrapper
+          model={formDataPath}
+          inModal={inModal}
+          viewDomain={viewDomain}
+          handleSubmit={(formData) => handleSubmit(
+            formData,
+            taxonomy
+          )}
+          handleSubmitFail={handleSubmitFail}
+          handleCancel={() => handleCancel(typeId)}
+          handleUpdate={handleUpdate}
+          onErrorDismiss={onErrorDismiss}
+          onServerErrorDismiss={onServerErrorDismiss}
+          scrollContainer={this.scrollContainer.current}
+          fields={dataReady && { // isManager, taxonomies,
+            header: {
+              main: this.getHeaderMainFields(),
+              aside: this.getHeaderAsideFields(
+                taxonomy,
+                parentOptions,
+                parentTaxonomy,
+              ),
+            },
+            body: {
+              main: this.getBodyMainFields(),
+              aside: this.getBodyAsideFields(),
+            },
+          }}
+        />
+      </Content>
+    );
+  }
+}
+
+CategoryNewForm.propTypes = {
+  loadEntitiesIfNeeded: PropTypes.func,
+  redirectIfNotPermitted: PropTypes.func,
+  handleSubmitRemote: PropTypes.func.isRequired,
+  handleSubmitFail: PropTypes.func.isRequired,
+  handleSubmit: PropTypes.func.isRequired,
+  handleCancel: PropTypes.func.isRequired,
+  handleUpdate: PropTypes.func.isRequired,
+  dataReady: PropTypes.bool,
+  authReady: PropTypes.bool,
+  viewDomain: PropTypes.object,
+  taxonomy: PropTypes.object,
+  parentOptions: PropTypes.object,
+  parentTaxonomy: PropTypes.object,
+  initialiseForm: PropTypes.func,
+  onErrorDismiss: PropTypes.func.isRequired,
+  onServerErrorDismiss: PropTypes.func.isRequired,
+  typeId: PropTypes.string,
+  formDataPath: PropTypes.string,
+  inModal: PropTypes.bool,
+};
+
+CategoryNewForm.contextTypes = {
+  intl: PropTypes.object.isRequired,
+};
+
+const mapStateToProps = (state, { typeId }) => ({
+  isAdmin: selectIsUserAdmin(state),
+  dataReady: selectReady(state, { path: DEPENDENCIES }),
+  authReady: selectReadyForAuthCheck(state),
+  taxonomy: selectTaxonomy(state, typeId),
+  parentOptions: selectParentOptions(state, typeId),
+  parentTaxonomy: selectParentTaxonomy(state, typeId),
+});
+
+function mapDispatchToProps(
+  dispatch,
+  {
+    formDataPath,
+    inModal,
+    onSaveSuccess,
+    onCancel,
+  }
+) {
+  return {
+    initialiseForm: (formData) => {
+      dispatch(formActions.reset(formDataPath));
+      dispatch(formActions.change(formDataPath, formData));
+    },
+    loadEntitiesIfNeeded: () => {
+      DEPENDENCIES.forEach((path) => dispatch(loadEntitiesIfNeeded(path)));
+    },
+    redirectIfNotPermitted: () => {
+      dispatch(redirectIfNotPermitted(USER_ROLES.MANAGER.value));
+    },
+    onErrorDismiss: () => {
+      dispatch(submitInvalid(true));
+    },
+    onServerErrorDismiss: () => {
+      dispatch(saveErrorDismiss());
+    },
+    handleSubmitFail: () => {
+      dispatch(submitInvalid(false));
+    },
+    handleSubmitRemote: (model) => {
+      dispatch(formActions.submit(model));
+    },
+    // handleSubmit: (formData, actions, actorsByActortype, taxonomy) => {
+    handleSubmit: (formData, taxonomy) => {
+      let saveData = formData.setIn(['attributes', 'taxonomy_id'], taxonomy.get('id'));
+      const formCategoryIds = getCheckedValuesFromOptions(formData.get('associatedCategory'));
+      if (List.isList(formCategoryIds) && formCategoryIds.size) {
+        saveData = saveData.setIn(['attributes', 'parent_id'], formCategoryIds.first());
+      } else {
+        saveData = saveData.setIn(['attributes', 'parent_id'], null);
+      }
+      dispatch(
+        newEntity({
+          path: API.CATEGORIES,
+          entity: saveData.toJS(),
+          redirect: !inModal ? ROUTES.CATEGORY : null,
+          onSuccess: inModal && onSaveSuccess,
+        })
+      );
+    },
+    handleCancel: (typeId) => {
+      if (inModal && onCancel) {
+        // cleanup
+        dispatch(submitInvalid(true));
+        dispatch(saveErrorDismiss());
+        onCancel();
+      } else {
+        dispatch(updatePath(`${ROUTES.TAXONOMIES}/${typeId}`, { replace: true }));
+      }
+    },
+    handleUpdate: (formData) => {
+      dispatch(updateEntityForm(formData));
+    },
+    onCreateOption: (args) => {
+      dispatch(openNewEntityModal(args));
+    },
+  };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(CategoryNewForm);
