--- app/containers/CategoryNew/actions.js
@@ -1,16 +1,0 @@
-/*
- *
- * CategoryNew actions
- *
- */
-
-import {
-  SAVE,
-} from './constants';
-
-export function save(data) {
-  return {
-    type: SAVE,
-    data,
-  };
-}
