+++ app/containers/EntityNew/constants.js
@@ -6,26 +6,5 @@ export const FORM_INITIAL = fromJS({
     draft: true,
     private: false,
     is_archive: false,
-    measuretype_id: '',
-    actortype_id: '',
-    resourcetype_id: '',
-    taxonomy_id: '',
-    manager_id: '',
-    parent_id: '',
-    title: '',
-    code: '',
-    description: '',
-    comment: '',
-    status: '',
-    url: '',
-    date_start: '',
-    date_end: '',
-    date_comment: '',
-    publication_date: '',
-    access_date: '',
-    target_comment: '',
-    status_comment: '',
-    activity_summary: '',
-    short_title: '',
   },
 });
