--- app/containers/IndicatorNew/actions.js
@@ -1,16 +1,0 @@
-/*
- *
- * IndicatorNew actions
- *
- */
-
-import {
-  SAVE,
-} from './constants';
-
-export function save(data) {
-  return {
-    type: SAVE,
-    data,
-  };
-}
