--- app/containers/PageNew/actions.js
@@ -1,16 +1,0 @@
-/*
- *
- * PageNew actions
- *
- */
-
-import {
-  SAVE,
-} from './constants';
-
-export function save(data) {
-  return {
-    type: SAVE,
-    data,
-  };
-}
