+++ app/containers/PageNew/index.js
@@ -8,227 +8,43 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 import Helmet from 'react-helmet';
-import { actions as formActions } from 'react-redux-form/immutable';
-
-import {
-  getTitleFormField,
-  getMenuTitleFormField,
-  getMarkdownFormField,
-  getStatusField,
-  getMenuOrderFormField,
-} from 'utils/forms';
-
-import { scrollToTop } from 'utils/scroll-to-component';
-import { hasNewErrorNEW } from 'utils/entity-form';
-
-import { CONTENT_SINGLE } from 'containers/App/constants';
-import { ROUTES, USER_ROLES } from 'themes/config';
-
-import {
-  loadEntitiesIfNeeded,
-  redirectIfNotPermitted,
-  updatePath,
-  updateEntityForm,
-  submitInvalid,
-  saveErrorDismiss,
-} from 'containers/App/actions';
-import {
-  selectReady,
-  selectReadyForAuthCheck,
-} from 'containers/App/selectors';
-
-import Content from 'components/Content';
-import ContentHeader from 'components/ContentHeader';
-import FormWrapper from './FormWrapper';
-
-import { selectDomainPage } from './selectors';
+import { intlShape, injectIntl } from 'react-intl';
 
 import messages from './messages';
-import { save } from './actions';
-import { DEPENDENCIES, FORM_INITIAL } from './constants';
 
-export class PageNew extends React.PureComponent { // eslint-disable-line react/prefer-stateless-function
-  constructor(props) {
-    super(props);
-    this.scrollContainer = React.createRef();
-  }
+import PageNewForm from './PageNewForm';
+import { selectDomain } from './selectors';
 
-  UNSAFE_componentWillMount() {
-    this.props.loadEntitiesIfNeeded();
-    this.props.initialiseForm('pageNew.form.data', FORM_INITIAL);
-  }
-
-  UNSAFE_componentWillReceiveProps(nextProps) {
-    // reload entities if invalidated
-    if (!nextProps.dataReady) {
-      this.props.loadEntitiesIfNeeded();
-    }
-    if (nextProps.authReady && !this.props.authReady) {
-      this.props.redirectIfNotPermitted();
-    }
-    if (hasNewErrorNEW(nextProps, this.props) && this.scrollContainer) {
-      scrollToTop(this.scrollContainer.current);
-    }
-  }
-
-  getHeaderMainFields = () => {
-    const { intl } = this.context;
-    return ([ // fieldGroups
-      { // fieldGroup
-        fields: [
-          getTitleFormField(intl.formatMessage),
-          getMenuTitleFormField(intl.formatMessage),
-          getMenuOrderFormField(intl.formatMessage),
-        ],
-      },
-    ]);
-  }
-
-  getHeaderAsideFields = () => {
-    const { intl } = this.context;
-    return ([{
-      fields: [
-        getStatusField(intl.formatMessage),
-        getStatusField(intl.formatMessage, 'private'),
-      ],
-    }]);
-  };
-
-  getBodyMainFields = () => {
-    const { intl } = this.context;
-    return ([{
-      fields: [getMarkdownFormField(intl.formatMessage, true, 'content')],
-    }]);
-  };
-
-  render() {
-    const { intl } = this.context;
-    const {
-      viewDomainPage,
-      dataReady,
-      onErrorDismiss,
-      onServerErrorDismiss,
-      handleCancel,
-      handleSubmitRemote,
-      handleSubmit,
-      handleSubmitFail,
-      handleUpdate,
-    } = this.props;
-    const { saveSending } = viewDomainPage.toJS();
-
-    return (
-      <div>
-        <Helmet
-          title={`${intl.formatMessage(messages.pageTitle)}`}
-          meta={[
-            {
-              name: 'description',
-              content: intl.formatMessage(messages.metaDescription),
-            },
-          ]}
-        />
-        <Content ref={this.scrollContainer}>
-          <ContentHeader
-            title={intl.formatMessage(messages.pageTitle)}
-            type={CONTENT_SINGLE}
-            icon="categories"
-            buttons={
-              dataReady ? [{
-                type: 'cancel',
-                onClick: this.props.handleCancel,
-              },
-              {
-                type: 'save',
-                disabled: saveSending,
-                onClick: () => handleSubmitRemote('pageNew.form.data'),
-              }] : null
-            }
-          />
-          <FormWrapper
-            model="pageNew.form.data"
-            handleSubmit={(formData) => handleSubmit(formData)}
-            handleSubmitFail={handleSubmitFail}
-            handleCancel={handleCancel}
-            handleUpdate={handleUpdate}
-            onErrorDismiss={onErrorDismiss}
-            onServerErrorDismiss={onServerErrorDismiss}
-            fields={dataReady && {
-              header: {
-                main: this.getHeaderMainFields(),
-                aside: this.getHeaderAsideFields(),
-              },
-              body: {
-                main: this.getBodyMainFields(),
-              },
-            }}
-            scrollContainer={this.scrollContainer.current}
-          />
-        </Content>
-      </div>
-    );
-  }
+export function PageNew({
+  viewDomain,
+  intl,
+}) {
+  return (
+    <div>
+      <Helmet
+        title={`${intl.formatMessage(messages.pageTitle)}`}
+        meta={[
+          {
+            name: 'description',
+            content: intl.formatMessage(messages.metaDescription),
+          },
+        ]}
+      />
+      <PageNewForm
+        viewDomain={viewDomain}
+        formDataPath="pageNew.form.data"
+      />
+    </div>
+  );
 }
 
 PageNew.propTypes = {
-  loadEntitiesIfNeeded: PropTypes.func.isRequired,
-  redirectIfNotPermitted: PropTypes.func,
-  handleSubmitRemote: PropTypes.func.isRequired,
-  handleSubmitFail: PropTypes.func.isRequired,
-  handleSubmit: PropTypes.func.isRequired,
-  handleCancel: PropTypes.func.isRequired,
-  handleUpdate: PropTypes.func.isRequired,
-  onErrorDismiss: PropTypes.func.isRequired,
-  onServerErrorDismiss: PropTypes.func.isRequired,
-  viewDomainPage: PropTypes.object,
-  dataReady: PropTypes.bool,
-  authReady: PropTypes.bool,
-  initialiseForm: PropTypes.func,
-};
-
-PageNew.contextTypes = {
-  intl: PropTypes.object.isRequired,
+  viewDomain: PropTypes.object,
+  intl: intlShape.isRequired,
 };
 
 const mapStateToProps = (state) => ({
-  viewDomainPage: selectDomainPage(state),
-  dataReady: selectReady(state, { path: DEPENDENCIES }),
-  authReady: selectReadyForAuthCheck(state),
+  viewDomain: selectDomain(state),
 });
 
-function mapDispatchToProps(dispatch) {
-  return {
-    initialiseForm: (model, formData) => {
-      dispatch(formActions.reset(model));
-      dispatch(formActions.change(model, formData, { silent: true }));
-    },
-    loadEntitiesIfNeeded: () => {
-      DEPENDENCIES.forEach((path) => dispatch(loadEntitiesIfNeeded(path)));
-    },
-    redirectIfNotPermitted: () => {
-      dispatch(redirectIfNotPermitted(USER_ROLES.ADMIN.value));
-    },
-    onErrorDismiss: () => {
-      dispatch(submitInvalid(true));
-    },
-    onServerErrorDismiss: () => {
-      dispatch(saveErrorDismiss());
-    },
-    handleSubmitFail: () => {
-      dispatch(submitInvalid(false));
-    },
-    handleSubmitRemote: (model) => {
-      dispatch(formActions.submit(model));
-    },
-    handleSubmit: (formData) => {
-      dispatch(save(formData.toJS()));
-    },
-    handleCancel: () => {
-      dispatch(updatePath(ROUTES.PAGES, { replace: true }));
-    },
-    handleUpdate: (formData) => {
-      dispatch(updateEntityForm(formData));
-    },
-  };
-}
-
-export default connect(mapStateToProps, mapDispatchToProps)(PageNew);
+export default connect(mapStateToProps, null)(injectIntl(PageNew));
