+++ app/containers/PageNew/selectors.js
@@ -4,8 +4,3 @@ export const selectDomain = createSelector(
   (state) => state.get('pageNew'),
   (substate) => substate
 );
-
-export const selectDomainPage = createSelector(
-  (state) => state.getIn(['pageNew', 'page']),
-  (substate) => substate
-);
