--- app/containers/ResourceNew/actions.js
@@ -1,16 +1,0 @@
-/*
- *
- * ActorNew actions
- *
- */
-
-import {
-  SAVE,
-} from './constants';
-
-export function save(data) {
-  return {
-    type: SAVE,
-    data,
-  };
-}
