+++ app/containers/ResourceNew/selectors.js
@@ -14,10 +14,6 @@ export const selectDomain = createSelector(
   (state) => state.get('resourceNew'),
   (substate) => substate
 );
-export const selectDomainPage = createSelector(
-  (state) => state.getIn(['resourceNew', 'page']),
-  (substate) => substate
-);
 
 export const selectConnectedTaxonomies = createSelector(
   selectActionTaxonomies,
